Move widget store to registry for shadcn CLI installation
- Move files from lib/ to registry/widgets/
- Add widget-store to registry.json
- Update docs with CLI and manual installation tabs
