+++ registry/widgets/widget-store-context.tsx
