+++ registry/widgets/widget-store.ts
